Dynamic Routes, https://vueschool.io/lessons/nuxtjs-dynamic-routes.
Dynamic routes tut loads a single blog post from an array of blogs. Created a posts folder, and renamed post.vue component to _id.vue. The underscore is a required Nuxt convention to show it's a URL route parameter. The route parameter name follows the underscore. So the pages/posts/_id.vue component dynamically filters through the blog array and will match URL parameter (e.g., "shimp" from localhost:3000/posts/shimp) with an id that equals shimp in the array. This data is returned in the template.